Cleaned up the other recipe registration
